Add files via upload
I tweaked a few things. Still have to fix the modal up, re-design the form a little bit, add an option on each of the library to remove them, and ideally learn a bit about localStorage and use that to save what a user has already added to their library.